Move heatmap endpoint from volunteer to submission view
